fetch_attachment: Fetch download URL with authorization header
This is currently failing with the Console error:

    Access to fetch at
    'https://chat.google.com/api/get_attachment_url?[...] from origin
    'chrome-extension://[...]' has been blocked by CORS policy: Response
    to preflight request doesn't pass access control check: Redirect is
    not allowed for a preflight request.

The OPTIONS preflight requests responds 302 Found, returning a Location
header beginning with:

    https://accounts.google.com/ServiceLogin

and the response header:

    access-control-request-headers: authorization

So the API request sends the Authorization header and token, but the
preflight request doesn't, and the download API doesn't like that.

But I've just discovered that OPTIONS requests must not send user
credentials as part of the spec:

* https://stackoverflow.com/questions/52072001/the-authorization-header-is-not-sent-in-options-call-by-the-browser
* https://stackoverflow.com/questions/40722700/add-authentication-to-options-request

Not sure what to do now.