fetch_attachment: Download the attachment using the Google Chat API
I had originally tried to download the attachment using the
`downloadUri` field from the chat message, but ran into a CORS error
because that link requires authentication even on OPTIONS requests.

That made me worried that I was going to have to download the file to
the file system using the chrome.downloads API and then use something
like https://web.dev/articles/origin-private-file-system to get the
bytes of the file and upload it to Google Drive.

Fortunately, after some reflection and a browse of the Google Chat API
reference, it turns out there's an API to get the attachment bytes. Easy
peasy.

Though the API docs make no mention of an "alt" query parameter, and the
API response without it is:

    {
      "error": {
        "code": 400,
        "message": "Invalid value for query parameter 'alt'. It must be set to \"media\".",
        "status": "INVALID_ARGUMENT"
      }
    }