content.js: Ideas to try to get top window's href from frame
Going to have to abandon this idea because we can't get around the
cross-origin problems. The top window's origin is "mail.google.com" and
the chat frame is at "chat.google.com".

We'd have to set up messaging to the background to get the current URL,
all for a verification that we're in the right chat frame. Deciding that
I'm not going to bother verifying against the top window's URL, that
should be simpler.