initialize_attachment_buttons: Experiment with mutation observer
I want a mutation observer that will trigger when new messages are
added, so we can inject a new attachment button when a new attachment is
present.

Had some trouble with the old `chat_container`, as only child nodes
would show up as `mutation.target`s. This meant I wouldn't get immediate
access to new `[data-topic-id]` nodes and would have to querySelector
for them. I decided to change the `chat_container` selector to select
the immediate parent of the `[data-topic-id]` nodes so I would be able
to get those in the mutation child list.

Still trying to get this working, but it's a step in the right
direction.