Disable open button when clicked, re-enable on success
I couldn't get `sendMessage` to work, so ended up using
`chrome.tabs.sendMessage` manually.

Need to pass the `group_id` around in order to be able to properly
select the button element.

Converted the injected "open" element into a button as that makes more
sense and we can disable it.