Cache Google Drive file IDs
After uploading a Google Chat attachment to Google Drive, cache its
Drive ID so we don't have to re-download and upload the attachment the
next time the attachment is opened with the extension.

Limited the maximum number of cached files to 100 to account for
chrome.storage space requirements. Once the limit is reached, old files
will need to be re-downloaded and recreated in Google Drive.