open_file_in_google_docs: Support 'document' and 'presentation' types
Support the other two office file types in addition to spreadsheets. Use
the file extension to determine which Google Docs program to use to open
the file.

Add an additional cache field containing the file's type to allow us to
select the right URL path from a cached file ID in addition to a
freshly-created Drive file.